Added benchmarks for rsocket GW transport (#86)
* Added benchmarks for rsocket GW transport

+++ src/main/java/io/scalecube/gateway/examples/GreetingService.java
@@ -16,7 +16,7 @@ public interface GreetingService {
   Flux<String> many(String name);
 
   @ServiceMethod("manyStream")
-  Flux<Integer> manyStream(Integer cnt);
+  Flux<Long> manyStream(Long cnt);
 
   @ServiceMethod("failing/one")
   Mono<String> failingOne(String name);
